examples/python/requirements.txt
grpcio
googleapis-common-protos
